Move all the login from index to main-app
